FunctionHandler: Handle keep-alive connections
When a connection is meant to be kept alive, we don't actually need to
close it. SwiftNIO will close it when appropriate.

For cases where the client specifically requests a connection be closed
after a response is sent (i.e. when "Connection" header is "close") we
close the connection immediately after sending the response ;)